Add possible template for creating spotlight animation
One of the main interactive features of the app will be an animated
score player that plays an audio recording one ore more versions of a
visual score, and highlights parts of the score as they are played. For
that, we need a way for authors to define the position and timing of
those highlights for us to import and make interactive. This first
option uses Adobe Illustrator for markup, with a SVG or optional CSS
export for conversion to digital.