Add alt version
This version of the animation shows the animated spotlight as more
uniform in color and opacity. This happens by clearing the drawn shape
after each frame, rather than letting them pile up on one another as in
the current main version of the animation. The problem with this though
is that because the Cage and Tudor spotlights are drawn simultaneously
by two different functions, the `clear();` function has the effect of
erasing both before they’re visible. So, only one set (Cage or Tudor)
can be seen at a time. To make this work for both, the functions and
all the variables passed into them, would need to be combined.